+++ material/src/material.ts
@@ -1,0 +1,4 @@
+export * from './modules/dialogPopup.module';
+export * from './components/basicDialogPopup/types';
+export * from './plugins/popup/components';
+export * from './plugins/popup';