+++ src/plugins/positioner/components.ts
@@ -1,1 +1,2 @@
 export * from './basic/basicPositioner.component';
+export * from './no/noPositioner.component';