+++ src/plugins/positioner/index.ts
@@ -1,3 +1,4 @@
 export * from './positioner.interface';
 
 export * from './basic/basicPositioner.interface';
+export * from './no/noPositioner.interface';