+++ src/plugins/positioner/no/noPositioner.interface.ts
@@ -1,0 +1,13 @@
+import {Positioner, PositionerOptions} from "../positioner.interface";
+
+/**
+ * Public API for 'NoPositionerComponent'
+ */
+export interface NoPositioner extends Positioner
+{}
+
+/**
+ * No positioner options
+ */
+export interface NoPositionerOptions extends PositionerOptions
+{}